% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control_mec}
\alias{control_mec}
\title{Title}
\usage{
control_mec(
  comp_rate = 0.9,
  deduplication = FALSE,
  add_xy = TRUE,
  maxit = 100,
  eps = 0.001,
  theta_est = "1",
  target_flr = 0.05,
  treshold = 1,
  increase_rate = 4,
  theta_start = 0.5
)
}
\arguments{
\item{comp_rate}{- a}

\item{deduplication}{- a}

\item{add_xy}{- a}

\item{maxit}{- a}

\item{eps}{- a}

\item{theta_est}{- a}

\item{target_flr}{- a}

\item{treshold}{- a}

\item{increase_rate}{- a}

\item{theta_start}{- a}
}
\description{
Title
}
