% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1}
\alias{example1}
\alias{example2}
\title{Collaboration in Research and Methodology for Official Statistics data.}
\format{
A data frame with 3,044 rows and 8 columns:
\describe{
  \item{firstname}{}
  \item{lastname}{}
  \item{sex}{}
  \item{dobday}{}
  \item{dobmon}{}
  \item{dobyear}{}
  \item{postcode}{}
  ...
}
}
\source{
<https://ec.europa.eu/eurostat/cros/content/essnet-di-fictitious-data-ons-job-training-record-linkage_en>
}
\description{
Data
}
\keyword{datasets}
