% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupervisedMec.R, R/UnsupervisedMec.R
\name{mec}
\alias{mec}
\title{Title mec}
\usage{
mec(
  A,
  B,
  vars,
  g,
  blockvars = NULL,
  string_comparator = FALSE,
  error_rate = FALSE,
  control = control_mec()
)

mec(
  A,
  B,
  vars,
  g,
  blockvars = NULL,
  string_comparator = FALSE,
  error_rate = FALSE,
  control = control_mec()
)
}
\arguments{
\item{A}{- first dataset to be linked}

\item{B}{- second dataset to be linked}

\item{vars}{- character vector of variables that should be compared.}

\item{g}{- a}

\item{blockvars}{- the variables defining the blocks or strata for which all pairs of x and y will be generated.}

\item{string_comparator}{- a}

\item{error_rate}{- if `TRUE` estimation guided by errors rate is generated}

\item{control}{- a control to function for tunning set of parameters}
}
\description{
Title mec

Title mec
}
