% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupervisedMec.R
\name{mecSup}
\alias{mecSup}
\title{Supervised Maximum Entropy Classifier for record linkage}
\usage{
mecSup(
  A,
  B,
  Omega = NULL,
  vars,
  g,
  blockvars = NULL,
  control = control_mec(),
  prob_ratio = "1"
)
}
\arguments{
\item{A}{The first data set to be linked.}

\item{B}{The second data set to be linked.}

\item{Omega}{Training set containing pairs of comparison vectors.}

\item{vars}{`character` vector of variables that should be compared.}

\item{g}{The dummy variable for matched and non-matched pairs in the Omega set.}

\item{blockvars}{The variables defining the blocks or strata for which all pairs of x and y will be generated.}

\item{control}{A list indicating control parameters to use in fitting model.}

\item{prob_ratio}{Method for the probability ratio estimation.}
}
\description{
\code{mecSup} fits MEC record linkage for a supervised learning.
}
\references{
D. Lee, L. C. Zhang and J. K. Kim. Maximum entropy classification for record linkage (2022)
}
\seealso{
[mecRecordLinkage::mec()]
}
\author{
Łukasz Chrostowski
}
