% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control_mec}
\alias{control_mec}
\title{Control parameters Maximum entropy classification}
\usage{
control_mec(
  deduplication = FALSE,
  add_xy = TRUE,
  maxit = 100,
  eps = 0.001,
  theta_est = "1",
  eta_est = "1",
  target_flr = 0.05,
  treshold = 1,
  increase_rate = 4,
  theta_start = 0.5,
  fixed_method = "Newton",
  all = FALSE
)
}
\arguments{
\item{deduplication}{generate pairs from only x. Ignore y. This is useful for deduplication of x. Default is `FALSE`.}

\item{add_xy}{add x and y as attributes to the returned pairs. This makes calling some subsequent operations that need x and y. Default is `FALSE`.}

\item{maxit}{The number of maximum iteration for fitting model in the unsupervised learning. Default is `100`.}

\item{eps}{Tolerance for fitting algorithms by default \code{1e-3}.}

\item{theta_est}{formula for theta estimation. Default is `1`.}

\item{eta_est}{formula for eta estimation. Default is `1`.}

\item{target_flr}{The target error rate (false link rate) for the fitting. Default is `.05`.}

\item{treshold}{The threshold value for fitting the algorithm by the error rates. Default is `1`.}

\item{increase_rate}{The increasing value for fitting the algorithm by the error rates. Default is `4`.}

\item{theta_start}{starting value for the theta parametet. Default is `.5`.}

\item{fixed_method}{fixed point method}

\item{all}{- a}
}
\description{
Control parameters Maximum entropy classification
}
\author{
Łukasz Chrostowski

#' @description \code{control_mec} constructs a list with all necessary control parameters
for MEC record linkage.
}
