% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnsupervisedMec.R
\name{mec}
\alias{mec}
\title{Unsupervised Maximum Entropy Classifier for a record linkage}
\usage{
mec(
  A,
  B,
  vars,
  blockvars = NULL,
  error_rate = FALSE,
  string_comparator = list(reclin2::identical()),
  control = control_mec()
)
}
\arguments{
\item{A}{first data set to be linked.}

\item{B}{second data set to be linked.}

\item{vars}{character vector of variables that should be compared.}

\item{blockvars}{the variables defining the blocks or strata for which all pairs of x and y will be generated.}

\item{error_rate}{if `TRUE`, estimation is guided by errors rate is generated.}

\item{string_comparator}{named vector of functions for comparing pairs of records.}

\item{control}{A list indicating control parameters to use in fitting model.}
}
\description{
\code{mec} fits MEC record linkage for an unsupervised learning.
}
\references{
D. Lee, L. C. Zhang and J. K. Kim. Maximum entropy classification for record linkage (2022)
}
\author{
Łukasz Chrostowski
}
